some section edited
